plugin: do not ghost sink pad of source bin
Do not ghost the sink pad for the source bin to avoid linking an unused
internal sink pad, as this may break delayed linking attempts.

The second argument from gst_parse_bin_from_description enables an
automatic creation and linking of ghost pads. These ghost pads are
linked to an unlinked pad, if there is exactly one unlinked pad, inside
the bin.

In the QtGstPlayer, the bin is assumed to be a video source which has
only the source pad. Therefore, ghosting a potential sink pad is not
necessary.

Ghosting the sink pad causes problems, if the pipeline contains a
(video) demux element. Initially, the demux does not provide a source
pad. Therefore, the downstream element (for example a parser or decoder)
inside the bin are not linked to the source pad of the demuxer when the
bin is created. Therefore, the unlinked sink pad of the downstream
elements is found and linked to the ghost pad. Later, linking the demux
element with its downstream elements fails, because the downstream sink
pads are already linked.

To avoid this error, disable the automatic ghosting of the unlinked pads
and manually ghost only the src pad.

Signed-off-by: Sven Püschel <s.pueschel@pengutronix.de>